dedupe grade_choice scorer
